# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Madnezz API",
        "description": "API for Madnezz application",
        "contact": {
            "name": "API Support",
            "email": "support@madnezz.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/v1/organization/context": {
            "get": {
                "tags": [
                    "Contexto Organizacional"
                ],
                "summary": "Obter contexto organizacional do usuário atual",
                "description": "Recupera o contexto organizacional do usuário autenticado incluindo posição, unidade e informações de departamento",
                "operationId": "a7d0607182c6665743b43cc55bbff620",
                "responses": {
                    "200": {
                        "description": "Contexto organizacional recuperado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "organization_id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "612b29b1-06a4-4885-ab98-de099feb6ecd"
                                                },
                                                "organization_name": {
                                                    "type": "string",
                                                    "example": "Madnezz Corporation"
                                                },
                                                "organization_code": {
                                                    "type": "string",
                                                    "example": "MADNEZZ"
                                                },
                                                "position_level": {
                                                    "type": "string",
                                                    "enum": [
                                                        "go",
                                                        "gr",
                                                        "store_manager"
                                                    ],
                                                    "example": "go"
                                                },
                                                "organization_unit_id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "e1acb852-4489-43d3-847f-ec027d016ccf"
                                                },
                                                "organization_unit_name": {
                                                    "type": "string",
                                                    "example": "Madnezz Corporation"
                                                },
                                                "organization_unit_type": {
                                                    "type": "string",
                                                    "enum": [
                                                        "company",
                                                        "regional",
                                                        "store"
                                                    ],
                                                    "example": "company"
                                                },
                                                "departments": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "enum": [
                                                            "administrative",
                                                            "financial",
                                                            "marketing",
                                                            "operations",
                                                            "trade",
                                                            "macro"
                                                        ]
                                                    },
                                                    "example": [
                                                        "administrative",
                                                        "financial",
                                                        "marketing",
                                                        "operations",
                                                        "trade",
                                                        "macro"
                                                    ]
                                                },
                                                "position_id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "a9da6ad7-8cf2-4f0a-a31a-b5ff49c39c8c"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Usuário não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário não autenticado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Usuário não possui posição ativa em nenhuma organização",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário não possui posição ativa em nenhuma organização"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/organization/dashboard": {
            "get": {
                "tags": [
                    "Acesso Hierárquico - Nível GO"
                ],
                "summary": "Acessar dashboard organizacional nível GO",
                "description": "Dashboard organizacional acessível apenas por usuários nível GO (Diretor)",
                "operationId": "e4d363a6abbc1ad1408ad7b05926b7f6",
                "responses": {
                    "200": {
                        "description": "Dashboard GO acessado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dashboard GO - visão geral da organização"
                                        },
                                        "context": {
                                            "description": "Contexto organizacional do usuário",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Nível hierárquico insuficiente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Nível hierárquico insuficiente. Necessário: go"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/regional/dashboard": {
            "get": {
                "tags": [
                    "Acesso Hierárquico - Nível GR"
                ],
                "summary": "Acessar dashboard regional nível GR",
                "description": "Dashboard regional acessível por usuários nível GR (Gerente Regional) e superiores",
                "operationId": "35b5301b4c6ecfbb8abfa194c53a956a",
                "responses": {
                    "200": {
                        "description": "Dashboard regional acessado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dashboard regional - visão geral da região"
                                        },
                                        "context": {
                                            "description": "Contexto organizacional do usuário",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Nível hierárquico insuficiente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Nível hierárquico insuficiente. Necessário: gr"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/store/dashboard": {
            "get": {
                "tags": [
                    "Acesso Hierárquico - Nível Loja"
                ],
                "summary": "Acessar dashboard da loja",
                "description": "Dashboard da loja acessível por todos os níveis hierárquicos (Gerente de Loja e superiores)",
                "operationId": "25940bb51f92e9916d289bf5b1edfa73",
                "responses": {
                    "200": {
                        "description": "Dashboard da loja acessado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dashboard da loja - dados específicos da loja"
                                        },
                                        "context": {
                                            "description": "Contexto organizacional do usuário",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Nível hierárquico insuficiente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Nível hierárquico insuficiente. Necessário: store_manager"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}